polaris_refactored/mock_external_system/requirements.txt
# Mock External System Dependencies

# Core dependencies (asyncio is part of Python standard library)
pyyaml>=6.0

# Testing dependencies
pytest>=7.0
pytest-asyncio>=0.21.0
hypothesis>=6.0

# Development dependencies
black>=23.0
mypy>=1.0
